Add Support for $ Commands; Improve Messaging
This all still needs lots of work, but it at least compiles.

- Parse $ commands and send them to a system function.
- Change signalOK to reportStatusMessage to make more robust
- Add settings library, define eeprom locations, define settings struct,
handle saving and loading and upgrading of eeprom data
- Add concet of alarm states, need a lot of work, no mechanism for
locking the machine

Most of this was copied from Grbl.  http://github.com/grbl/grbl

A ton of this is still commented out and a lot of work is still needed
to get things working, but this will compile.

+++ cnc_ctrl_v1/Kinematics.h
@@ -37,7 +37,7 @@
             float D             = 1;                              //distance between sprocket centers
             float R             = 10.1;                                //sprocket radius
             float rotationDiskRadius = 0;                              //distance from the bit to the attachment point
-            int kinematicsType  = 1;                                   //Which version of the kinematics to use. 1 is quadrilateral, 2 is triangular
+            byte kinematicsType  = 1;                                   //Which version of the kinematics to use. 1 is quadrilateral, 2 is triangular
 
 
             //machine dimensions
